project/marcal-llm-lab/pinfoodmap/requirements.txt
selenium
transformers
torch
pandas
python-dotenv